requirements.txt
prometheus_client
requests
python-slugify